updated readmes and put brotein on listh
