+++ .eslintcache 2
@@ -1,0 +1,1 @@
+[{"/Users/jasontilley/Documents/Homework/react-portfolio/src/components/footer.js":"1","/Users/jasontilley/Documents/Homework/react-portfolio/src/components/about.js":"2","/Users/jasontilley/Documents/Homework/react-portfolio/src/components/contact.js":"3","/Users/jasontilley/Documents/Homework/react-portfolio/src/App.js":"4","/Users/jasontilley/Documents/Homework/react-portfolio/src/components/nav.js":"5","/Users/jasontilley/Documents/Homework/react-portfolio/src/index.js":"6","/Users/jasontilley/Documents/Homework/react-portfolio/src/components/appCard.js":"7","/Users/jasontilley/Documents/Homework/react-portfolio/src/components/portfolio.js":"8","/Users/jasontilley/Documents/Homework/react-portfolio/src/components/background.js":"9"},{"size":1474,"mtime":1606788446259,"results":"10","hashOfConfig":"11"},{"size":3314,"mtime":1606801058076,"results":"12","hashOfConfig":"11"},{"size":1614,"mtime":1606801774128,"results":"13","hashOfConfig":"11"},{"size":504,"mtime":1606795666394,"results":"14","hashOfConfig":"11"},{"size":2347,"mtime":1606791859267,"results":"15","hashOfConfig":"11"},{"size":263,"mtime":1606788446261,"results":"16","hashOfConfig":"11"},{"size":1462,"mtime":1606797765603,"results":"17","hashOfConfig":"11"},{"size":1197,"mtime":1606800735334,"results":"18","hashOfConfig":"11"},{"size":652,"mtime":1606801634756,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":"22"},"13226w0",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"39"},"/Users/jasontilley/Documents/Homework/react-portfolio/src/components/footer.js",["40","41"],"import React from \"react\";\nimport styled from \"styled-components\";\nimport { Row, Col } from \"reactstrap\";\nimport { FontAwesomeIcon } from '@fortawesome/react-fontawesome'\nimport { fab } from '@fortawesome/free-brands-svg-icons'\n\n\nconst FooterWrapper = styled.div`\n  span {\n    margin: 0 20px;\n    color: white;\n  }\n\n  svg {\n    font-size: 30pt;\n    color: white;\n  }\n\n  footer {\n    display: flex;\n    justify-content: center;\n    flex-direction: row;\n    align-items: flex-end;\n    padding: 30px 200px;\n    background-color: black;\n  }\n\n  @media screen and (max-width: 480px) {\n    span {\n      margin: 0 20px;\n      color: white;\n    }\n\n    icon {\n      font-size: 30pt;\n      color: white;\n    }\n\n    footer {\n      display: flex;\n      justify-content: center;\n      flex-direction: row;\n      align-items: flex-end;\n      padding: 30px 10px;\n      width: 100%\n    }\n  }\n`;\n\nfunction Footer() {\n  console.log('fab: ', fab);\n  const { faGithub, faLinkedin } = fab;\n  return (\n    <FooterWrapper>\n      <footer>\n        <span>  \n          <a href=\"https://www.linkedin.com/in/jason-t-tilley/\">  \n            <FontAwesomeIcon icon={faLinkedin} size=\"2x\" />\n          </a>\n        </span>\n        <span>\n          <a href=\"https://github.com/jttilley\">  \n            <FontAwesomeIcon icon={faGithub} size=\"2x\" />\n          </a>\n        </span>\n        <span>\n          &copy; Copyright 2020\n        </span>\n      </footer>\n    </FooterWrapper>\n  );\n}\n\nexport default Footer;","/Users/jasontilley/Documents/Homework/react-portfolio/src/components/about.js",[],"/Users/jasontilley/Documents/Homework/react-portfolio/src/components/contact.js",[],"/Users/jasontilley/Documents/Homework/react-portfolio/src/App.js",[],"/Users/jasontilley/Documents/Homework/react-portfolio/src/components/nav.js",["42"],"/Users/jasontilley/Documents/Homework/react-portfolio/src/index.js",[],"/Users/jasontilley/Documents/Homework/react-portfolio/src/components/appCard.js",[],"/Users/jasontilley/Documents/Homework/react-portfolio/src/components/portfolio.js",[],"/Users/jasontilley/Documents/Homework/react-portfolio/src/components/background.js",["43"],"import React, { Children } from \"react\"\nimport styled from \"styled-components\";\n\nconst FullBackgroundWrapper = styled.div`\n  .fullbackground {\n    background-image: url(\"https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/background.jpeg?raw=true\");\n    background-size: cover;\n    background-position: center;\n    height: 100%;\n  }\n  @media screen and (max-width: 480px) {\n    .fullbackground {\n      background-image: none;\n    }\n  }\n`;\n\nfunction Background({ children }) {\n  return(\n    <FullBackgroundWrapper>\n      <div className=\"fullbackground\">{children}</div>\n    </FullBackgroundWrapper>\n  );\n}\n\nexport default Background;",{"ruleId":"44","severity":1,"message":"45","line":3,"column":10,"nodeType":"46","messageId":"47","endLine":3,"endColumn":13},{"ruleId":"44","severity":1,"message":"48","line":3,"column":15,"nodeType":"46","messageId":"47","endLine":3,"endColumn":18},{"ruleId":"44","severity":1,"message":"49","line":1,"column":17,"nodeType":"46","messageId":"47","endLine":1,"endColumn":25},{"ruleId":"44","severity":1,"message":"50","line":1,"column":17,"nodeType":"46","messageId":"47","endLine":1,"endColumn":25},"no-unused-vars","'Row' is defined but never used.","Identifier","unusedVar","'Col' is defined but never used.","'useState' is defined but never used.","'Children' is defined but never used."]
