+++ src/App 2.js
@@ -1,6 +1,5 @@
-import logo from './logo.svg';
 import './App.css';
-import { BrowserRouter as Router, Route, Link, NavLink } from 'react-router-dom';
+import { HashRouter as Router, Route } from 'react-router-dom';
 import Portfolio from './components/portfolio';
 import About from './components/about';
 import Contact from './components/contact';
@@ -8,11 +7,10 @@ import Contact from './components/contact';
 function App() {
   return (
     <>
-      <Router>
+      <Router basename='/'>
         <Route exact path="/" component={Portfolio} />
         <Route exact path="/about" component={About} />
         <Route exact path="/contact" component={Contact} />
-      
       </Router>
     </>
   );
