+++ src/assets/data/portfolio-apps.json
@@ -3,49 +3,31 @@
     "title": "Weather Trips",
     "github": "https://github.com/jttilley/Weather-Trips",
     "deployed": "https://jttilley.github.io/Weather-Trips/",
-    "img": "https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/Weather-Trips.png?raw=true",
-    "imgDescription": "A mapped route with weather cards next to it."
+    "img": "../images/Weather-Trips.png"
   },{
     "title": "BillBoard Budget",
     "github": "https://github.com/jttilley/BillBoards",
     "deployed": "https://sleepy-eyrie-55325.herokuapp.com/",
-    "img": "https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/BillBoardBudget.png?raw=true",
-    "imgDescription": "Bill Board budget page"
-  },{
-    "title": "Employee Directory",
-    "github": "https://github.com/jttilley/employee-directory",
-    "deployed": "https://jttilley.github.io/employee-directory/",
-    "img": "https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/employeesDir.png?raw=true",
-    "imgDescription": "List of employees info and a search bar"
-  },{
-    "title": "Workout Tracker",
-    "github": "https://github.com/jttilley/WorkoutTracker",
-    "deployed": "https://workout-tracker-37.herokuapp.com/",
-    "img": "https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/Workout_Tracker.png?raw=true",
-    "imgDescription": "Charts and graphs about workouts"
+    "img": "../images/BillBoardBudget.png"
   },{
     "title": "Weather Dashboard",
     "github": "https://github.com/jttilley/WeatherDashboard",
     "deployed": "https://jttilley.github.io/WeatherDashboard/index.html",
-    "img": "https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/Weather-Dashboard.png?raw=true",
-    "imgDescription": "Weather cards with weather icons and info"
+    "img": "../images/Weather-Dashboard.png"
   },{
     "title": "Note Taker",
     "github": "https://github.com/jttilley/NoteTaker",
     "deployed": "https://secure-lowlands-99445.herokuapp.com/",
-    "img": "https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/NoteTaker.png?raw=true",
-    "imgDescription": "A list of note titles and an open note."
+    "img": "../images/NoteTaker.png"
   },{
     "title": "Work Day Planner",
     "github": "https://github.com/jttilley/WorkDayPlanner",
     "deployed": "https://jttilley.github.io/WorkDayPlanner/index.html",
-    "img": "https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/Work-Day-Planner.png?raw=true",
-    "imgDescription": "A work day broken up by hour color coded to show what has passed, where we are and what is to come."
+    "img": "../images/Work-Day-Planner.png"
   },{
-    "title": "Eat Da Burger",
-    "github": "https://github.com/jttilley/burger",
-    "deployed": "https://cryptic-tor-81296.herokuapp.com/",
-    "img": "https://github.com/jttilley/react-portfolio/blob/gh-pages/public/images/burger.png?raw=true",
-    "imgDescription": "2 lists, one with burgers to be devoured and one with burgers that have been devoured"
+    "title": "Brotien",
+    "github": "https://github.com/jttilley/Brotein",
+    "deployed": "https://brotein.herokuapp.com/",
+    "img": "../images/brotein.jpg"
   }
 ]
